tarjeta bigint

+++ data/creaciontablasyesquema.sql
@@ -171,7 +171,7 @@ CREATE TABLE LIL_MIX.funcionalidadxrol ( rol_id INT,
 					 FOREIGN KEY (funcionalidad_id) REFERENCES LIL_MIX.funcionalidad(funcionalidad_id)
 					 )
 
-CREATE TABLE LIL_MIX.tarjeta ( 	tarjeta_numero INT NOT NULL PRIMARY KEY,
+CREATE TABLE LIL_MIX.tarjeta ( 	tarjeta_numero BIGINT NOT NULL PRIMARY KEY,
 				tarjeta_tipo VARCHAR(30),
 				tarjeta_fecha_vencimiento DATETIME,
 				tarjeta_id_cliente INT
@@ -186,7 +186,7 @@ CREATE TABLE LIL_MIX.cargaDeCredito ( carga_id INT NOT NULL IDENTITY(1,1) PRIMAR
 				      carga_id_cliente INT NOT NULL FOREIGN KEY REFERENCES LIL_MIX.cliente(cliente_id),
 				      carga_tipo_de_pago INT NOT NULL FOREIGN KEY REFERENCES LIL_MIX.tipoDePago(tipo_de_pago_id), --TODAVIA NO EJECUTAMOS
 				      carga_monto BIGINT NOT NULL,
-				      carga_tarjeta_numero INT FOREIGN KEY REFERENCES LIL_MIX.tarjeta(tarjeta_numero) 
+				      carga_tarjeta_numero BIGINT FOREIGN KEY REFERENCES LIL_MIX.tarjeta(tarjeta_numero) 
 				     )
 
 CREATE TABLE LIL_MIX.semestre ( semestre_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
